requirements.txt
Django==1.10.1
gunicorn==19.4.5
dj-database-url==0.4.1
whitenoise==3.2.1
djangorestframework
markdown
django-filter
generic==0.3.1
django_chartit
django-markup-deprecated
dj-static
sphinx
xlwt
drfdocs
Pillow